Python3 compatible fix
