+++ src/pyforce/__init__.py
@@ -1,14 +1,27 @@
 '''pyforce is there to access the RESTful Force.com API'''
+
 import logging
-from xmlclient import _tPartnerNS, _tSObjectNS, _tSoapNS, SoapFaultError,\
-    SessionTimeoutError, DEFAULT_SERVER_URL
-from xmlclient import Client as XMLClient
-from pyforce import Client as PythonClient
+
+from pyforce.xmlclient import (
+    _tPartnerNS,
+    _tSObjectNS,
+    _tSoapNS,
+    DEFAULT_SERVER_URL,
+    SoapFaultError,
+    SessionTimeoutError
+)
+from pyforce.xmlclient import Client as XMLClient
+from pyforce.pyforce import Client as PythonClient
 
 __all__ = (
-    'XMLClient', '_tPartnerNS', '_tSObjectNS', '_tSoapNS', 'tests',
-    'SoapFaultError', 'SessionTimeoutError', 'PythonClient',
-    'DEFAULT_SERVER_URL'
+    'XMLClient',
+    '_tPartnerNS',
+    '_tSObjectNS',
+    '_tSoapNS',
+    'DEFAULT_SERVER_URL',
+    'PythonClient',
+    'SoapFaultError',
+    'SessionTimeoutError'
 )
