+++ src/pyforce/xmltramp.py
@@ -1,9 +1,9 @@
-"""xmltramp: Make XML documents easily accessible."""
 
-from xml.sax.handler import EntityResolver, DTDHandler, ContentHandler,\
-    ErrorHandler
 from xml.sax import make_parser
-from xml.sax.handler import feature_namespaces
+from xml.sax.handler import (
+    EntityResolver, DTDHandler, ContentHandler,
+    ErrorHandler, feature_namespaces
+)
 
 __version__ = "2.18"
 __author__ = "Aaron Swartz"
@@ -11,14 +11,9 @@ __credits__ = "Many thanks to pjz, bitsko, and DanC."
 __copyright__ = "(C) 2003-2006 Aaron Swartz. GNU GPL 2."
 
 
-if not hasattr(__builtins__, 'True'):
-    True = 1
-    False = 0
-
-
 def isstr(mystring):
     '''Check if string is a string or unicode'''
-    return isinstance(mystring, str) or isinstance(mystring, unicode)
+    return isinstance(mystring, str)
 
 
 def islst(myitem):
@@ -145,7 +140,7 @@ class Element(object):
     def __unicode__(self):
         text = ''
         for x in self._dir:
-            text += unicode(x)
+            text += x
         return ' '.join(text.split())
 
     def __str__(self):
@@ -337,143 +332,10 @@ def seed(fileobj):
 
 
 def parse(text):
-    from StringIO import StringIO
+    from io import StringIO
     return seed(StringIO(text))
 
 
 def load(url):
     import urllib
     return seed(urllib.urlopen(url))
-
-
-def unittest():
-    parse('<doc>a<baz>f<b>o</b>ob<b>a</b>r</baz>a</doc>').__repr__(1, 1) == \
-        '<doc>\n\ta<baz>\n\t\tf<b>o</b>ob<b>a</b>r\n\t</baz>a\n</doc>'
-
-    assert str(parse("<doc />")) == ""
-    assert str(parse("<doc>I <b>love</b> you.</doc>")) == "I love you."
-    assert parse("<doc>\nmom\nwow\n</doc>")[0].strip() == "mom\nwow"
-    assert str(
-        parse('<bing>  <bang> <bong>center</bong> </bang>  </bing>')
-    ) == "center"
-    assert str(parse('<doc>\xcf\x80</doc>')) == '\xcf\x80'
-
-    d = Element(
-        'foo',
-        attrs={'foo': 'bar'},
-        children=['hit with a', Element('bar'), Element('bar')]
-    )
-
-    try:
-        d._doesnotexist
-        raise "ExpectedError", "but found success. Damn."
-    except AttributeError:
-        pass
-    assert d.bar._name == 'bar'
-    try:
-        d.doesnotexist
-        raise "ExpectedError", "but found success. Damn."
-    except AttributeError:
-        pass
-
-    assert hasattr(d, 'bar') is True
-
-    assert d('foo') == 'bar'
-    d(silly='yes')
-    assert d('silly') == 'yes'
-    assert d() == d._attrs
-
-    assert d[0] == 'hit with a'
-    d[0] = 'ice cream'
-    assert d[0] == 'ice cream'
-    del d[0]
-    assert d[0]._name == "bar"
-    assert len(d[:]) == len(d._dir)
-    assert len(d[1:]) == len(d._dir) - 1
-    assert len(d['bar',]) == 2
-    d['bar',] = 'baz'
-    assert len(d['bar',]) == 3
-    assert d['bar']._name == 'bar'
-
-    d = Element('foo')
-
-    doc = Namespace("http://example.org/bar")
-    bbc = Namespace("http://example.org/bbc")
-    dc = Namespace("http://purl.org/dc/elements/1.1/")
-    d = parse("""<doc version="2.7182818284590451"
-        xmlns="http://example.org/bar"
-        xmlns:dc="http://purl.org/dc/elements/1.1/"
-        xmlns:bbc="http://example.org/bbc">
-        <author>John Polk and John Palfrey</author>
-        <dc:creator>John Polk</dc:creator>
-        <dc:creator>John Palfrey</dc:creator>
-        <bbc:show bbc:station="4">Buffy</bbc:show>
-    </doc>""")
-
-    assert repr(d) == '<doc version="2.7182818284590451">...</doc>'
-    assert d.__repr__(1) == (
-        '<doc xmlns="http://example.org/bar" xmlns:bbc="http://example.org/bbc"'
-        ' xmlns:dc="http://purl.org/dc/elements/1.1/" version="2.71828182845904'
-        '51"><author>John Polk and John Palfrey</author><dc:creator>John Polk</'
-        'dc:creator><dc:creator>John Palfrey</dc:creator><bbc:show bbc:station='
-        '"4">Buffy</bbc:show></doc>'
-    )
-    assert d.__repr__(1,1) == (
-        '<doc xmlns="http://example.org/bar" xmlns:bbc="http://example.org/bbc"'
-        ' xmlns:dc="http://purl.org/dc/elements/1.1/" version="2.71828182845904'
-        '51">\n\t<author>John Polk and John Palfrey</author>\n\t<dc:creator>Joh'
-        'n Polk</dc:creator>\n\t<dc:creator>John Palfrey</dc:creator>\n\t<bbc:s'
-        'how bbc:station="4">Buffy</bbc:show>\n</doc>'
-    )
-
-    assert repr(parse("<doc xml:lang='en' />")) == '<doc xml:lang="en"></doc>'
-
-    assert str(d.author) == str(d['author']) == "John Polk and John Palfrey"
-    assert d.author._name == doc.author
-    assert str(d[dc.creator]) == "John Polk"
-    assert d[dc.creator]._name == dc.creator
-    assert str(d[dc.creator,][1]) == "John Palfrey"
-    d[dc.creator] = "Me!!!"
-    assert str(d[dc.creator]) == "Me!!!"
-    assert len(d[dc.creator,]) == 1
-    d[dc.creator,] = "You!!!"
-    assert len(d[dc.creator,]) == 2
-
-    assert d[bbc.show](bbc.station) == "4"
-    d[bbc.show](bbc.station, "5")
-    assert d[bbc.show](bbc.station) == "5"
-
-    e = Element('e')
-    e.c = '<img src="foo">'
-    assert e.__repr__(1) == '<e><c>&lt;img src="foo"></c></e>'
-    e.c = '2 > 4'
-    assert e.__repr__(1) == '<e><c>2 > 4</c></e>'
-    e.c = 'CDATA sections are <em>closed</em> with ]]>.'
-    assert e.__repr__(1) == (
-        '<e><c>CDATA sections are &lt;em>closed&lt;/em> with ]]&gt;.</c></e>'
-    )
-    e.c = parse(
-        '<div xmlns="http://www.w3.org/1999/xhtml">i<br /><span></span>love<br'
-        ' />you</div>'
-    )
-    assert e.__repr__(1) == (
-        '<e><c><div xmlns="http://www.w3.org/1999/xhtml">i<br /><span></span>'
-        'love<br />you</div></c></e>'
-    )
-
-    e = Element('e')
-    e('c', 'that "sucks"')
-    assert e.__repr__(1) == '<e c="that &quot;sucks&quot;"></e>'
-
-    assert quote("]]>") == "]]&gt;"
-    assert quote(
-        '< dkdkdsd dkd sksdksdfsd fsdfdsf]]> kfdfkg >'
-    ) == '&lt; dkdkdsd dkd sksdksdfsd fsdfdsf]]&gt; kfdfkg >'
-
-    assert parse('<x a="&lt;"></x>').__repr__(1) == '<x a="&lt;"></x>'
-    assert parse(
-        '<a xmlns="http://a"><b xmlns="http://b"/></a>'
-    ).__repr__(1) == '<a xmlns="http://a"><b xmlns="http://b"></b></a>'
-
-if __name__ == '__main__':
-    unittest()
