+++ tests/test_xmlclient.py
